wip: Fix tests
